Do not test new entries in the pile
When a new modular RPM gets added to the pile, if we test it again, it
will be tested only against default modules and bare RPMs. The correct
module dependencies will not be used.

Care is needed to not mark modules themselves as tested, since that
could cause some alternative dependency paths to not be processed.