Mark packages from disabled modules as not considered
If the module is disabled, we don't want to mask any bare/default RPMs
just because there is a package in that disabled module.

There are two changes here: the masking function ignores modular
packages that are not considered, and when we disable unavailable
module, we also disable all packages in it.