Treat * as input solvable specially
If the input specifies * as a solvable, it means the user wants to
include all packages in the repo. No masking should be done. However
if a package is provided in a module as well as a bare RPM, the best
solvable function will keep only one version.

The fix for this is to ignore all modular packages when a wildcard is
processed. The modular packages will still be included because their
module will be included. Only latest version of the bare RPM will get
in.