Add tests for repo loading functionality
We have a test suite for checking that the hybrid solving is working.
This suite bypasses the normal repo loading function by reading a
libsolv testtag file and a modules.yaml file directly.

So add two more tests to check that we can load local and remote
repositories. Hopefully this will catch mistakes like #68.

Signed-off-by: Rafael Fonseca <r4f4rfs@gmail.com>