Add a failing test case for build in multiple modules
There are two versions of the same module stream in the repo. They both
contain the exact same RPM.

Fus reports dependency issues with the modular package.

If the input modulemd only contains the 2020 version, the process
finishes fine. If only 2019 is present, the test fails because the older
version is pulled in.